Fix move component to ui folder
